github/scripts/requirements.txt
yt-dlp==2024.08.06
google-generativeai==0.8.1
supabase==2.7.4
resend==0.8.0
requests==2.32.3
